Add memory initialization to the prologue, and various improvemnts
The improvements are all dealing with the creation of the constant
table:

1. New assoc-i function allows to do assoc-like operations but can
search for the key in a specified column (instead of just the first one)
2. Due to 1, the ordering of the constants table/dictionary is now
improved: the address is in the first column.
3. A new function create-imm-prologue creates the memory initialization
code. The conversion of the constant table to a memory image string is
also done here.
4. Perhaps various minor changes